+++ src/Entity/Media.php
@@ -16,9 +16,9 @@ use Ixocreate\Entity\DefinitionCollection;
 use Ixocreate\Entity\EntityInterface;
 use Ixocreate\Entity\EntityTrait;
 use Ixocreate\Media\MediaInterface;
-use Ixocreate\Type\Entity\DateTimeType;
-use Ixocreate\Type\Entity\UuidType;
-use Ixocreate\Type\TypeInterface;
+use Ixocreate\Schema\Type\DateTimeType;
+use Ixocreate\Schema\Type\TypeInterface;
+use Ixocreate\Schema\Type\UuidType;
 
 final class Media implements EntityInterface, DatabaseEntityInterface, MediaInterface
 {