+++ src/Schema/Type/VideoType.php
@@ -7,22 +7,22 @@
 
 declare(strict_types=1);
 
-namespace Ixocreate\Media\Type;
+namespace Ixocreate\Media\Schema\Type;
 
 use Doctrine\DBAL\Types\GuidType;
-use Ixocreate\Entity\Type\AbstractType;
-use Ixocreate\Entity\Type\Type;
 use Ixocreate\Media\Config\MediaConfig;
 use Ixocreate\Media\Entity\Media;
+use Ixocreate\Media\Schema\Element\VideoElement;
 use Ixocreate\Media\Uri\MediaUri;
-use Ixocreate\Schema\BuilderInterface;
-use Ixocreate\Schema\ElementInterface;
-use Ixocreate\Schema\ElementProviderInterface;
-use Ixocreate\Schema\Elements\DocumentElement;
-use Ixocreate\Type\DatabaseTypeInterface;
-use Ixocreate\Type\TypeInterface;
+use Ixocreate\Schema\Builder\BuilderInterface;
+use Ixocreate\Schema\Element\ElementInterface;
+use Ixocreate\Schema\Element\ElementProviderInterface;
+use Ixocreate\Schema\Type\AbstractType;
+use Ixocreate\Schema\Type\DatabaseTypeInterface;
+use Ixocreate\Schema\Type\Type;
+use Ixocreate\Schema\Type\TypeInterface;
 
-final class DocumentType extends AbstractType implements DatabaseTypeInterface, ElementProviderInterface, \Serializable
+final class VideoType extends AbstractType implements DatabaseTypeInterface, ElementProviderInterface, \Serializable
 {
     /**
      * @var MediaType
@@ -57,7 +57,7 @@ final class DocumentType extends AbstractType implements DatabaseTypeInterface, 
 
         if (!empty($mediaType->value()) && \in_array(
             $mediaType->value()->mimeType(),
-            $this->mediaConfig->documentWhitelist()
+            $this->mediaConfig->videoWhitelist()
         )) {
             $type->mediaType = $mediaType;
         }
@@ -132,12 +132,12 @@ final class DocumentType extends AbstractType implements DatabaseTypeInterface, 
 
     public static function serviceName(): string
     {
-        return 'document';
+        return 'video';
     }
 
     public function provideElement(BuilderInterface $builder): ElementInterface
     {
-        return $builder->get(DocumentElement::class);
+        return $builder->get(VideoElement::class);
     }
 
     /**
@@ -155,8 +155,8 @@ final class DocumentType extends AbstractType implements DatabaseTypeInterface, 
      */
     public function unserialize($serialized)
     {
-        /** @var DocumentType $type */
-        $type = Type::get(DocumentType::serviceName());
+        /** @var VideoType $type */
+        $type = Type::get(VideoType::serviceName());
         $this->uri = $type->uri;
         $this->mediaConfig = $type->mediaConfig;