+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## [Unreleased](https://github.com/ixocreate/media-package/compare/0.2.1...develop)
 ### Changed
 - Upgrade Publish config to Application v0.2
+### Fixed
+- fix -c flag in RecreateImageDefinition
 
 ## [v0.2.1 (2019-05-03)](https://github.com/ixocreate/media-package/compare/0.2.0...0.2.1)